+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added square-root scaling as an additional usage chart scale option.
 
+### Changed
+
+- Show the package version in the modal header.
+
 ## [1.2.1] - 2026-08-12
 
 ### Changed